chore: migrated z3 initialiseContext and getZ3Context to separate ffile and exported
